Skeleton set up, ready to make module
Set up a base page that I want to replicate for the three pages of my
website, and the CSS to make that happen. From here I will create a js
module that will be imported into my main file, allowing me to append
the data for each page to be loaded as the buttons are clicked.