Finished Restaurant page project
Added some styling to the page, including a feature that highlights
the button of the page that is currently active on the site, so the
user can easily tell what page they are on. I realized that I didn't do
the project exactly how I was supposed to from a modules perspective,
so I made the highlighting functions into their own module so I could
practice that a little bit more.